True ND block arrays, in progress
This sets the principle that a block array and all its blocks must have the same
number of dims, changing deeply how block arrays are built from blocks. Among
other things, this enforces some dimension consistency and allows for some
broadcasting of the blocks.

Support for general ndim case is improved.

Broadcasting of shapes is implemented, but broadcasting of the arrays themselves
is still missing.