fix(fit_eval): do not return fitted model
This was intended to make exploration of cv result richer, but is a catastrophic
memory killer if you have ten grid points, ten fold cv, and each model contains
a copy of the data. Under galp, indexing can be used to never deserialize the
fitted model again, but it requires some care ; and cv is also used outside of
galp where eager serialization is not available.

If this feature is ever re-introduced, it needs to come with a way to disable
it. For now no use case has been found so we just remove it.